+++ components/ui/button/index.vue
@@ -1,39 +1,42 @@
 <template>
-  <button
+  <component
+    :is="component"
     class="outline-none focus:outline-none hover:outline-none cursor-pointer relative disabled:border-opacity-60 disabled:cursor-not-allowed"
     :disabled="disabled"
     type="submit"
     :class="{
       'disabled:bg-gray-100 disabled:text-gray-400':
-        type == 'primary' || type == 'secondary',
+        type === 'primary' || type === 'secondary',
       'disabled:border-gray-400 disabled:text-gray-400 disabled:bg-transparent':
         type.includes('outline'),
-      'bg-primary hover:bg-primary-dark  text-white': type == 'primary',
+      'bg-primary hover:bg-primary-dark  text-white': type === 'primary',
       'lg:px-7 px-5': padding,
-      'bg-secondary text-primary': type == 'secondary',
+      'bg-secondary text-primary': type === 'secondary',
       'text-primary border border-solid border-primary bg-transparent hover:bg-primary hover:text-white':
-        type == 'outline-primary',
+        type === 'outline-primary',
       'text-secondary border border-solid border-secondary bg-transparent hover:bg-secondary hover:text-gray-900':
-        type == 'outline-secondary',
+        type === 'outline-secondary',
       'text-white border border-solid border-white bg-transparent hover:bg-white hover:text-primary':
-        type == 'outline-white',
+        type === 'outline-white',
       'text-gray-400 border border-solid border-gray-400 bg-transparent hover:bg-gray-500 hover:text-gray-200':
-        type == 'outline-gray',
-      'bg-transparent text-primary': type == 'link',
+        type === 'outline-gray',
+      'bg-transparent text-primary': type === 'link',
       'rounded-full': rounded,
       'community-button': communityStyles,
     }"
     :padding="[padding]"
     :margin="[margin]"
     :style="!disabled ? styles : null"
+    :target="target"
+    :href="link"
     v-on="inputListeners"
   >
-    <slot />
-  </button>
+    <slot/>
+  </component>
 </template>
 
 <script>
-import { mapGetters } from 'vuex'
+import {mapGetters} from 'vuex'
 
 export default {
   name: 'Button',
@@ -104,9 +107,7 @@ export default {
           click(event) {
             if (!vm.link) {
               vm.$emit('click', event)
-              return
             }
-            vm.navigate()
           },
         }
       )
@@ -134,12 +135,9 @@ export default {
         ...(this.customStyle || {}),
       }
     },
-  },
-  methods: {
-    navigate() {
-      if (!this.link) return
-      window.open(this.link, this.target || undefined)
-    },
+    component() {
+      return this.link ? 'a' : 'button'
+    }
   },
 }
 </script>